Implemented AABB Collision Detection

+++ starlight/starlight/core/MemMgr/MemMgr.cpp
@@ -3,14 +3,13 @@
 // Init our MemMgr singleton to null
 std::unique_ptr<MemMgr> MemMgr::memMgr(nullptr);
 
-
 MemMgr::MemMgr(uint totalSpace)
 	: regionSize(totalSpace),
 	start(AllocSpace(regionSize)),
-	frameData(new DoubleBufferAllocator(totalSpace * 0.1)),
-	poolData(new PoolAllocator(totalSpace * 0.2)),
-	globalData(new StackAllocator(totalSpace * 0.1)),
-	levelData(new StackAllocator(totalSpace * 0.6))
+	frameData(new DoubleBufferAllocator(static_cast<unsigned int>(totalSpace * 0.1))),
+	poolData(new PoolAllocator(static_cast<unsigned int>(totalSpace * 0.2))),
+	globalData(new StackAllocator(static_cast<unsigned int>(totalSpace * 0.1))),
+	levelData(new StackAllocator(static_cast<unsigned int>(totalSpace * 0.6)))
 {
 	// Log("allocated block of size: " << regionSize << "at address: " << start);
 }
